Extract construction API module
